% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_car_data.R
\name{get_car_data}
\alias{get_car_data}
\title{get_car_data}
\usage{
get_car_data(make, model)
}
\arguments{
\item{make}{The make according to the catalog}

\item{model}{The model according to the catalog}
}
\value{
A list containig the url image, price, version prices and competitors
}
\description{
Obtain data from Autocosmos according to the selecte make and model
}
\examples{
make <- get_make()[1]
model <- get_model(make)[1]
get_car_data(make, model)
}
\author{
Jean Arreola
}

