% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_make.R
\name{get_make}
\alias{get_make}
\title{get_make}
\usage{
get_make()
}
\value{
A character string containing all the makes in the catalog.
}
\description{
Obtain all the makes included in Autocosmos website for the current year
}
\examples{
get_make()
}
\author{
Jean Arreola
}

