% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{get_model}
\usage{
get_model(make)
}
\value{
A character string containing all the models in the catalog.
}
\description{
Obtain all the possible
}
\examples{
make <- get_make()[1]
get_model(make)
}
\author{
Jean Arreola
}

