% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_car_data.R
\name{get_car_data}
\alias{get_car_data}
\title{get_car_data}
\usage{
get_car_data(brand, submarca)
}
\arguments{
\item{brand}{The brand according to the catalog}

\item{make}{The make according to the catalog}
}
\value{
A list containig the url image, price, model prices and competitors
}
\description{
Obtain data from Autocosmos according to the selecte brand and make
}
\examples{
get_car_data( "Volkswagen", "Jetta")
}
\author{
Jean Arreola
}

